pos/requirements/common.txt
Django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
# Pillow==7
django-cors-headers==3.2.1
django-debug-toolbar==3.2.1
django-modeltranslation==0.17.3
pyfcm==1.5.3
requests==2.26.0
django-import-export==2.0.2
apscheduler==3.7.0
sendgrid==6.9.2
python-gettext==4.0
# psycopg2==2.8.6
django-filter==2.4.0
boto3==1.18.21
pdfkit==0.6.1
python-decouple==3.6.0
drf-yasg==1.20.0
django-jazzmin==2.5.0